master Use bat for manpager; Use exa instead of ls
alias.fish will now check to see if exa (ls alternative) exists on the
system. If found, it will use aliases to use exa instead of `ls`.

config.fish will now check to see if bat (cat alternative) exists on the
system. If found, it will use bat instead of `less` as manpager.